Updated shr package.
git-svn-id: svn://daria.forty-two.fr/shr/trunk@342 a7dca5fb-7094-41a2-a069-28cb1271ccbd

+++ openembedded/packages/images/shr-image.bb
@@ -177,10 +177,10 @@ shr_rootfs_postprocess() {
 
 
     # setup opkg config
-    rm ./etc/opkg/all-feed.conf
-    rm ./etc/opkg/armv4t-feed.conf
-    rm ./etc/opkg/fic-gta02-feed.conf
-    rm ./etc/opkg/neo1973-feed.conf
+    if [ -a ./etc/opkg/all-feed.conf ]; then rm ./etc/opkg/all-feed.conf; fi
+    if [ -a ./etc/opkg/armv4t-feed.conf ]; then rm ./etc/opkg/armv4t-feed.conf; fi
+    if [ -a ./etc/opkg/fic-gta02-feed.conf ]; then rm ./etc/opkg/fic-gta02-feed.conf; fi
+    if [ -a ./etc/opkg/neo1973-feed.conf ]; then rm ./etc/opkg/neo1973-feed.conf; fi
     if [ -a ./etc/opkg/om-gta01-feed.conf ]; then rm ./etc/opkg/om-gta02-feed.conf; fi
     if [ -a ./etc/opkg/om-gta02-feed.conf ]; then rm ./etc/opkg/om-gta02-feed.conf; fi
